:sparkles: rework structure of gopher-learn
